refactor: move head and worker to grpc
