+++ jina/serve/runtimes/__init__.py
@@ -1,19 +1,3 @@
-def list_all_runtimes():
-    """List all public runtimes that can be used directly with :class:`jina.orchestrate.pods.Pod`
-
-    # noqa: DAR101
-    # noqa: DAR201
-    """
-    from jina.serve.runtimes.base import BaseRuntime
-    from jina.serve.runtimes.worker import WorkerRuntime
-
-    return [
-        k
-        for k, s in locals().items()
-        if isinstance(s, type) and issubclass(s, BaseRuntime)
-    ]
-
-
 def get_runtime(name: str):
     """Get a public runtime by its name
 
@@ -22,8 +6,8 @@ def get_runtime(name: str):
     """
     from jina.serve.runtimes.base import BaseRuntime
     from jina.serve.runtimes.gateway import GatewayRuntime
-    from jina.serve.runtimes.head import HeadRuntime
-    from jina.serve.runtimes.worker import WorkerRuntime
+    from jina.serve.runtimes.head.grpc import HeadRuntime
+    from jina.serve.runtimes.worker.grpc import WorkerRuntime
 
     s = locals()[name]
     if isinstance(s, type) and issubclass(s, BaseRuntime):
